minor restructuring and documentation update
