--- src/adafruit/core/forecast/CONFIG.properties
@@ -1,8 +1,0 @@
-[ConnectionData]
-APIKeyDomain=https://home.openweathermap.org/api_keys
-#APIKeyName=<enter your token name here - optional>
-#APIKey=<enter your token here>
-[ApplicationData]
-#CityID=<enter your city ID here if you want to define it statically>
-#CityName=<enter your city name here if you want to define it statically>
-#Country=<enter your country here if you want to define it statically>
