+++ src/test/adafruit/forecast/neopixel_base_test.py
@@ -29,8 +29,8 @@ Potential color ranges are:
 @deffield    updated: Updated
 '''
 from adafruit.core.neopixel_base import NeoPixelBase
-from adafruit.core.forecast.forecast_colors import ForecastNeoPixelColors
 from adafruit.core.util.cmd_functions import cmd_options
+from adafruit.controller.forecast.forecast_colors import ForecastNeoPixelColors
 
 class NeoPixelForecastColorsTest(NeoPixelBase):
