+++ src/test/adafruit/forecast/neopixel_base_threadtest.py
@@ -32,12 +32,12 @@ Potential color ranges are:
 @deffield    updated: Updated
 '''
 from adafruit.core.neopixel_base import NeoPixelBase
-from adafruit.core.forecast.forecast_colors import ForecastNeoPixelColors
 from adafruit.core.util.cmd_functions import cmd_options
 import board
 import neopixel
 import time
 from threading import Thread
+from adafruit.controller.forecast.forecast_colors import ForecastNeoPixelColors
 
 # defines the time in seconds that it will take to from an illuminated to faded state
 # calculation will be based on 190 individual steps for dimming from which 100 are twice the speed - step divider 280
